autocompletion code wasn't updated for the new context switch code
git-svn-id: svn://svn.code.sf.net/p/bluefish/code/trunk@3968 7640d186-872a-4c71-b7b2-a5696852cc35

+++ testcode/bftestview2/bftextview2_autocomp.c
@@ -269,7 +269,7 @@ static void print_ac_items(GCompletion *gc) {
 void autocomp_run(BluefishTextView *btv, gboolean user_requested) {
 	GtkTextIter cursorpos,iter;
 	GtkTextBuffer *buffer;
-	guint16 contextnum;
+	gint contextnum;
 	gunichar uc;
 	buffer = gtk_text_view_get_buffer(GTK_TEXT_VIEW(btv));
